refactor: 에러 핸들러 추가 및 service 계층 클래스 분리
